feat: checkout-mvc

+++ Shopping_ver1/Models/CartItemModel.cs
@@ -2,7 +2,7 @@
 {
     public class CartItemModel
     {
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
         public string ProductName { get; set; }
         public int Quantity { get; set; }
         public decimal Price { get; set; }
@@ -11,10 +11,7 @@
             get { return Quantity * Price; }
         }
         public string Image { get; set; }
-        public CartItemModel()
-        {
-
-        }
+        public CartItemModel() { }
         public CartItemModel(ProductModel product)
         {
             ProductId = product.Id;
